Move around some shared intents printing code
