+++ src/cmd/networkmapper/list/mapper-list.go
@@ -2,6 +2,7 @@ package list
 
 import (
 	"context"
+	"github.com/otterize/otterize-cli/src/pkg/intentsprinter"
 	"github.com/otterize/otterize-cli/src/pkg/mapperclient"
 	"github.com/otterize/otterize-cli/src/pkg/output"
 	"github.com/spf13/cobra"
@@ -24,22 +25,12 @@ var ListCmd = &cobra.Command{
 			if err != nil {
 				return err
 			}
-			for _, service := range servicesIntents {
-				output.PrintStdout("%s in namespace %s calls:", service.Client.Name, service.Client.Namespace)
-				for _, intent := range service.Intents {
-					if len(intent.Namespace) != 0 {
-						output.PrintStdout("  - %s in namespace %s", intent.Name, intent.Namespace)
-						continue
-					}
-
-					output.PrintStdout("  - %s", intent.Name)
-				}
-
-			}
-
 			if len(servicesIntents) == 0 {
 				output.PrintStderr("No connections found. The network mapper detects (1) connections that are currently open and (2) DNS lookups while a connection is being initiated, for connections between pods on this cluster.")
+			} else {
+				intentsprinter.ListFormattedIntents(intentsprinter.MapperIntentsToAPIIntents(servicesIntents))
 			}
+
 			return nil
 		})
 	},
